Removes TODO's around handling dirty pages
This commit removes the TODO's that were present earlier around handling of dirty pages for autovacuum page relocation. However, we don't actually need journalling to implement that at all. Now, the cache movements are correctly implemented by removing the dirty status for the destination page and keeping the dirty status of the source page