implement RowSet functions
They are used for example for things like DELETE RETURNING.
SQLite uses a complex data structure: it's either a vector for small
sets, or a set of trees split by ranges for larger sets.

I don't think we should do that unless absolutely necessary.
In this commit, I am implementing *only* the simple version so we can
move forward. I will then implement some users of the instruction
to make sure that it is working well, and then we can come back to make
a decision on the data structure.