Implement a btree clear operation
We already have a btree destroy operation, that fully destroys a btree.
SQLite also has a clear operation, that is used in statements like
DELETE FROM T; without qualifiers.

It will drop the btree contents, count how many rows were affected, and
reinitialize the root so we can start using the table again.

I am implementing this by sharing as much code as possible with the
destroy operation, only adding a processed cells count. That isn't very
expensive, so it is okay if destroy does that too.

After the btree is destroyed, then we reinitialize a new page.