Basic implementation of savepoints
Only savepoint and rollback to are properly implemented, another caveat
is that it must have a opened wal and autocommit enabled. So it don't
work inside a BEGIN statement.

The major problem is how to represent the modified pages in-memory when
not relying purely on wal. SQLite uses an in-memory rollback journal,
even in wal mode, to keep track of modified page, but we unlikely would
want to do this.