fix for join with the USING keyword
USING is a convenient way to express an equijoin. Instead of writing
a.id = b.id, you can just JOIN USING (a), and the database will find a
table with that name in both tables.

It is not hard to support this with our operator, since it is ultimately
an equijoin, but we have to do the work to deduplicate the columns when
producing the final schema.

Fixes #3362