core/storage: Layer MVCC on top of the B-Tree
We want a hybrid storage architecture where we don't bring everything to
memory. Hekaton solves this problem with Siberia, which separates row
versions between hot (memory) and cold (disk) storage. To determine if a
row is in cold storage, Siberia has a bloom filter to reduce cold storage
accesses. However, the performance characteristics for us are different:
we already have the page cache so looking up cold storage is not a
problem. Furthermore, pager read transactions guarantee that we have
snapshot in place, so we can just layer MVCC index on top of the b-tree,
never inserting row versions on reads, only on writes.