Fix premature EOF bug in SortedChunk::read()
When the sorter's internal read buffer was full and a prefetch read was
issued, the read would request 0 bytes. This obviously returned 0 bytes read,
and was incorrectly interpreted as EOF, causing the sorter to think the chunk
was exhausted when most data remained unread.

This resulted in "Corrupt database: Incomplete record" errors.

Fix:

- Avoid issuing pointless 0 byte reads
- Distinguish between "no room in buffer" and "nothing left to read",
  and only the latter causes EOF