fix/btree: disable move_to_rightmost optimization with triggers
We have an optimization in the btree where if:

- We want to reach the rightmost leaf page, and
- We know the rightmost page and are already on it

Then  we can skip a seek.

The problem is this optimization should NEVER be used in cases where we cannot be sure that the btree wasn't modified from under us e.g. by a trigger subprogram.

Hence, disable it when we are executing a parent program that has triggers which
will fire.

No AI was used for this PR.